+++ lib/frontend/components/marble-form/index.js
@@ -5,8 +5,8 @@ import _ from 'lodash'
 import moment from 'moment'
 import Select, { Creatable } from 'react-select'
 import DatePicker from 'react-datepicker'
-import RadioWidget from './RadioWidget'
 
+import RadioWidget from './radio-widget'
 import FileWidget from './file-widget'
 
 class EmailWidget extends Component {
@@ -714,6 +714,7 @@ class MarbleForm extends Component {
   }
 
   render () {
+    const {buttonClassName} = this.props
     let {
       errorMessage,
       successMessage,
@@ -746,7 +747,7 @@ class MarbleForm extends Component {
 
     let formActions = this.props.children
     if (!formActions) {
-      const className = classNames('button is-primary is-fullwidth', {
+      const className = classNames(buttonClassName || 'button is-primary is-fullwidth', {
         'is-loading': loading
       })
